Experiment: Keep fully realized vdom around between renders
To render provided hiccup, we must "parse" it: separate id and class name from
tag name, convert the tag name keyword to a string, and find the attrs map.
Since this is done when the hiccup is rendered the first time, it should not be
necessary to re-parse "old" on subsequent renders.

This change builds up a fully realized vdom structure and keeps it around for
future renders. This reduces the number of "parse" calls by roughly half (e.g.
the number of old nodes) for every render call except the first - in return for
some overhead in constructing the full data structure, and roughly double the
memory usage for tracking the old vdom, as we need both the original hiccup for
equality checks, and the structured vdom for comparing keys and tag names.