Use "significant nils"
The hiccup replicant renders will commonly be generated by code. If that code
uses conditionals like `when`, a `nil` is significant as it says something about
the structure. A `nil` in `old` means where there used to be nothing there is
now a new node. A `nil` in `new` means where there used to be a node, there
should now be nothing. This change causes less element re-creation for nodes
without keys.