Allow aliases to return aliases
With this commit, aliases enter the vdom tree. What? Yes. Previously, aliases
where treated as alternative expressions of hiccup. The problem with that
approach is that each hiccup node can only be represented by one alias
expression. This model breaks down when an alias returns another alias in the
root position.

With this commit, aliases are no longer stored as alternative representations of
hiccup nodes. Instead, aliases are part of the VDOM tree. When the diffing
algorithm encounters an alias, it will resolve it while still holding on to the
DOM node in question.