Pass dispatch function to wrapped event handlers
This feature allows users to opt-in to having their event handler wrapped such
that it is passed the same map that the global dispatch function receives.
Via this map, the event handler receives the global dispatch function.

This feature makes it possible to build aliases that take event handler data,
and have them dispatched from custom events in third party libraries, etc.